generic hashtable implementation
The original hashtable is modified to accept any value type. The key is
still uint64 because atomic is impossible for general key type, and it
seems that we don't really need the key to be generic.
A "better" hash function is added to replace the original identity hash
function, which should hopefully give less collision, seems doing
nothing for now because our test cases don't have much collision.