Make tests optional + fix option declarations
Break out the MeshIO from the tests, and make building the tests
optional. In cases where building/re-building the tests is not
wanted/needed, this can shorten build times by over 25% on my machine.
A major motivator for this is building 3rd party bindings.

Also included here are descriptors for the option declarations, which I
noticed do not actually set the default value as expected if the
description argument is omitted (defaults to OFF, rather than ON for
_TEST and _PYBIND for example -- on my machine anyway).