Fix test-{cmake,pkgconfig}.sh to create, not append, to test files.
The test.cpp file and the CMakeLists.txt/Makefile were created with
the '>>' append in shell instead of '>'. Which means these files were
growing on multiple invocations if *-consumer was not deleted
between calls (which means that the test would fail the second time
around).

While at it, add the manifold/parallel.h include also to the
pkg-config test.

Make all includes 'system-header' style with angle brackets, as this
is what users of the library would use.