Use a bundled Clipper2 copy instead of downloading
This preserves the ability to disable cross_section (and thus the use of
clipper) but replaces the Download-and-Build fetch with a bundled source
copy to avoid the network connection.  If we go this route, we're also
building Clipper2 directly into Manifold so client libraries don't have
to worry about linking it in.