Need to fix compilation flag lost in CMake refactor
Looks like we lost the O3 flag always being on.  This was good for some
fun trying to figure out why performance had dropped off so dramatically
- turned out it was simply because prior builds had NEVER been without
O3, so all builds - Debug or Release - were optimized.

On the other hand, there are arguments for and against O3 with debug
builds.  The optimizations complicate/obscure debugging, but it's also
an order of magnitude slower running.  Rather than be completely either
or, this PR defines a MANIFOLD_OPTIMIZED variable that a user can
specify to ON explicitly.  Build types Release and RelWithDebInfo will
always have O3, and for other cases it'll be up to whether the user has
explicitly requested optimized building with MANIFOLD_OPTIMIZED.